Adds SearchBuilder module to mixin to Blacklight catalog search builder class.
Adds Catalog module to mixin to Blacklight catalog controller.
